refactor of getLinearValueFromOptions
